[BE] feat: passport TS 적용 #95
+ strategy 파일 네이밍 컨벤션 변경
