--- backend/passport/index.js
@@ -1,7 +1,0 @@
-const jwt = require('./jwt-strategy');
-const local = require('./local-strategy');
-
-module.exports = (passport) => {
-  local(passport);
-  jwt(passport);
-};
